fix https://github.com/radiasoft/devops/issues/170
Collect issue and review comments with iterator
pkjson.dump_pretty now so issues are readable.
Note that labels and milestones are returned in full in each issue
so don't need to collect those separately.
Doesn't get releases (not in use by RadiaSoft)
store each issue in a separate file and tar the result