WIP radiasoft/sirepo#3732: Update pkresource to search additional packages
Sirepo needs to be able to load resources that live outside of its namespace.
This updates pkresource to support searching for resources beyond the
caller_context / caller_module.

The additional packages are searched first so resources found in them can
override the resources in Sirepo (or any caller context/module).