Fix #418: remove argh pin
In argh v0.30.0 there was a breaking change in the way argh mapped
function arguments to CLI arguments. For example, if you had a
function like `def f(x, bar=None)` prior to v0.30.0 bar was an
optional keyword argument which means on the CLI it was set with a
flag (--bar BAR). In v0.30.0 and up bar became a positional arg with a
default. So, you could no longer invoke the function from the CLI with
--bar set.

To fix this we now set the NameMappingPolicy to BY_NAME_IF_HAS_DEFAULT
 which retains the closest approximation with the old behavior.