Fix #529: Don't keep reference to tornado.httpclient.AsyncHTTPClient (#530)
By keeping a reference to AsyncHTTPClient in self._client it causes
sockets opened by the client to never be freed. Eventually, we
run out of file descriptors. Writing the code this way w/o
self._client prevents that from happening. This is a stop gap
until we can fix the underlying bug.